+++ src/config/majors.ts
